Fix lib3mf for libzip >= 1.9.0
Latest upload of libzip 1.11.1 introduces an incompatibility with lib3mf,
all attempts to open a file fail with ZIP_ER_OPNOTSUPP. This is because
lib3mf does not implement the ZIP_SOURCE_FREE command in its callback; this
was not enforced previously.

Fix by back-porting a patch from lib3fm v2 that adds support for the missing
command.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>